Refactored geo-indexing routines.
